sign in page
Just a side project to practice if else statements and using functions in js. Going to add some more CSS and make everything responsive. Also going to add some javascript dom to make it more complex. Just started this today and made a lot of progress. This should be done by tomorrow. Glade I am getting the hang of if and else if statements with javascript dom. This will be helpful for my current project I am working on that I will continue after this little side project.