Finished!!!!
Added some if else statements for the password and added a button so you can see and hide the password. Looked up how to center a div :( but other than that I've learned a lot about how to use the if else statements in javascript and put them in the HTML display. This was probably the most fun I had because I took what I learned with if else statements and then added them to javascript dom which I learn in my last project and put them together in HTML CSS and javascript and using functions. This isn't a complex project but for me, this is a milestone and I cant wait to call his project trash in the next week once I learn more js.